split invite member modal
